Remove function `callproc`
That function was problematic in multiple ways, but mostly because
evaluating its return value would lead to a call to `undefined` and
hence to an error. However, one would always have to define a type of
the resulting call to prevent type ambiguities. Long story short,
`callproc` was unsafe and not even very pleasant to work with.

If a lua function doesn't return a value (e.g. `print`), it can still be
called via `callfunc`. The return type should be `Result ()` in that
case.