Drop thread/coroutine related functions
Lua coroutines are implemented using longjmp, which plays badly with
haskell's RTS. Tests to get coroutines working from within haskell were
unsuccessful, so these functions are removed as to avoid confusion for
users of the library:

- newthread
- pushthread
- resume
- xmove
- yield

They can be added back if there is a way to make coroutines work or if
there are other good reasons to provide these wrapper functions.

The raw C bindings are unaffected by this step.