Wrap values received from stack with safePeek
A failed peek is not always an exceptional situation, returning an Error
object can be more natural than throwing an exception in such a case.
The `safePeek` function is added to the FromLuaStack typeclass, and all
instances implement that method instead of `peek`. The `peek` function
is now expressed in terms of `safePeek` and `throw`.