Use registry value instead of global as for errors
Haskell functions can signal errors by returning a special error object.
This was previously done using a global, since this allowed Lua scripts
to test for this value. However, having a global is weird, and error
translations are now done in C. Therefore, the special error object is
now stored in the registry field `HSLUA_ERR`.

The name change from `_HASKELLERR` to `HSLUA_ERR` is due to the fact
that Lua considers globals and registry fields starting with an
underscore as reserved. The manual suggests to use a string containing
the library name to avoid collisions.

Closes: #65