Let numeric retrieval functions return Maybes
Having a single function which returns a value wrapped in Maybe is
cleaner than having two functions, one of which can effectively be
expressed as `fmap (fromMaybe 0) . tonumber`.

A side-effect of this change is that `peek` will now also accept strings
which can be converted into the expected number type.