Add benchmark for `setraw`
The function `lua_setraw`, which the Haskell function `setraw` wraps, is
imported without extra C wrappings and as an *unsafe* C function by
default. Benchmarking it provides us with data which can be used to
estimate the overhead caused by the multiple levels of wrapping
introduced to make error-throwing functions safe to use.