Use alternative low-level error signaling
Errors are now detected in Haskell by passing a pointer to C functions
and checking the status after the function finishes. On the C side, the
pointer values are set to the result of the `lua_pcall` invocation.

This has the advantage of being cleaner than the previous method of
using negative return values to signal errors: error detection and
handling becomes more uniform; wrapper function signatures have the same
signature as the wrapped functions, except for the additional
`int *status` parameter; raw function bindings no longer depend on an
`Error` module.

The disadvantage is a (barely noticeable) impact on performance, as an
additional allocation is required for each call. However, calls to
wrapped functions require `safe` FFI imports, which makes calls very
expensive. One allocation carries no weight.