Parameterize main Lua type over the error type
The main type is renamed to `LuaE` and now takes the type of exceptions
as an additional argument. The type `Lua` is a specialization which uses
the default error handling mechanism.

Custom error handling can be used by making an exception type an
instance of the new typeclass `LuaError` which handles conversion
between Haskell Exception and a Lua error.