hslua-objectorientation: allow integers as aliases
Aliases can now be of type `AliasIndex`, so integers can now be defined
as aliases for other properties. Means the function `alias` now takes
`AliasIndex` instead of `Name`, and entails changes to the types
`UDTypeWithList`, `UDType`, and `Member`. Also, `AliasIndex` is made
into an instance of the Eq and Ord type classes.