lua: support arbitrary number of uservalues.
Lua 5.4 allows an arbitrary number of uservalues to be associated with
userdata objects. The functions `lua_newuserdata`, `lua_getuservalue`,
and `lua_setuservalue` are replaced with the new functions
`lua_newuserdatauv`, `lua_getiuservalue`, and `lua_setiuservalue`,
respectively.

The function `hslua_newhsuserdata` is renamed to `hslua_newhsuserdatauv`
and takes the number of associated uservalues as an additional argument.