hslua-packaging: add type initializers to Module record
This allows to associate types with a module. For performance reasons,
the types are not initialized when the module is pushed, but only when
they are first used. However, the documentation Lua object for each
module now has an additional field `types`. The new field contains a
function that returns the names of all associated types; it also
initializes these types.