hslua-core: add `setwarnf'` for simple warning handling
The built-in method of setting a warn function is flexible but not
straight-forward to use. The new `setwarnf'` function allows to set a
Haskell action as a warning hook: the default warning behavior is kept
in place, but, in addition to the default action, the given hook is
called on the concatenated warning messages.